Try adding pytest and move code to functions.py
