+++ requirements.txt
@@ -1,3 +1,4 @@
 schedule
 requests
 varname
+pytest
